requirements.txt
networkx>=3.2
numpy>=1.26
pandas>=2.2
PyYAML>=6.0.2
pycti>=6.8.0
requests>=2.32
Flask>=3.0
Flask-Cors>=4.0
python-dotenv>=1.0
pydantic>=2.11
pydantic-settings>=2.2
pytest>=8.2
